PSplines

Documentation for PSplines.

PSplines.D1matrix

PSplines.D2matrix

PSplines.D3matrix

PSplines.buildOm1

PSplines.buildOm2

PSplines.buildOm3

PSplines.D1matrix — Method

D1matrix(nvertices)

Returns the first difference of dimension nvercies-1 x nvertices. It is important to note that the numerical first derivative of a vector produces a vector with 1 component less than the input vector.

#Examples

julia> D1matrix(3)
2×3 Matrix{Float64}:
 -1.0   1.0  0.0
  0.0  -1.0  1.0

PSplines.D2matrix — Method

D2matrix(nvertices)

Returns the second difference of dimension nvercies-2 x nvertices. It is important to note that the numerical second derivative of a vector produces a vector with 2 component less than the input vector.

#Examples

julia> D2matrix(4)
2×4 Matrix{Float64}:
 1.0  -2.0   1.0  0.0
 0.0   1.0  -2.0  1.0

PSplines.D3matrix — Method

D3matrix(nvertices)

Returns the second difference of dimension nvercies-3 x nvertices. It is important to note that the numerical second derivative of a vector produces a vector with 2 component less than the input vector.

#Examples

julia> D3matrix(5)
2×5 Matrix{Float64}:
 -1.0   3.0  -3.0   1.0  0.0
  0.0  -1.0   3.0  -3.0  1.0

PSplines.buildOm1 — Method

buildOm1(nvertices)

Returns a matrix full of zeros with the appropiate dimensions to apply a selective penalization to the first differences vector. The default 0 value, implies no penalization on the first derivative.

#Examples

julia> buildOm1(5)
4×4 Matrix{Float64}:
 0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0

PSplines.buildOm2 — Method

buildOm2(nvertices)

Returns a matrix full of zeros with the appropiate dimensions to apply a selective penalization to the second differences vector. The default 0 value, implies no penalization on the second derivative. #Examples

julia> buildOm2(5)
3×3 Matrix{Float64}:
 0.0  0.0  0.0
 0.0  0.0  0.0
 0.0  0.0  0.0

PSplines.buildOm3 — Method

buildOm3(nvertices)

Returns a matrix full of zeros with the appropiate dimensions to apply a selective penalization to the third differences vector. The default 0 value, implies no penalization on the third derivative. #Examples

julia> buildOm3(5)
2×2 Matrix{Float64}:
 0.0  0.0
 0.0  0.0