PSplines

Documentation for PSplines.

PSplines.BSpline

PSplines.D1matrix

PSplines.D2matrix

PSplines.D3matrix

PSplines.buildKnots

PSplines.buildOm1

PSplines.buildOm2

PSplines.buildOm3

PSplines.buildW

PSplines.equiSpacedPosition

PSplines.evalBSplineBasis

PSplines.knots

PSplines.vertices

PSplines.BSpline — Type

BSpline{T}

bspline is a structures which includes the knots and the vertices of a BSpline.

PSplines.D1matrix — Method

D1matrix(nvertices)

Returns the first difference of dimension nvercies-1 x nvertices. It is important to note that the numerical first derivative of a vector produces a vector with 1 component less than the input vector.

#Examples

julia> D1matrix(3)
2×3 Matrix{Float64}:
 -1.0   1.0  0.0
  0.0  -1.0  1.0

PSplines.D2matrix — Method

D2matrix(nvertices)

Returns the second difference of dimension nvercies-2 x nvertices. It is important to note that the numerical second derivative of a vector produces a vector with 2 component less than the input vector.

#Examples

julia> D2matrix(4)
2×4 Matrix{Float64}:
 1.0  -2.0   1.0  0.0
 0.0   1.0  -2.0  1.0

PSplines.D3matrix — Method

D3matrix(nvertices)

Returns the second difference of dimension nvercies-3 x nvertices. It is important to note that the numerical second derivative of a vector produces a vector with 2 component less than the input vector.

#Examples

julia> D3matrix(5)
2×5 Matrix{Float64}:
 -1.0   3.0  -3.0   1.0  0.0
  0.0  -1.0   3.0  -3.0  1.0

PSplines.buildKnots — Method

buildKnots(xin, xfin, nvertices)

Returns the knots of a BSpline with beginning xin, end xfin with nvertices components. The knots returned by this function are equispaced between xin and xfin.

#Examples

julia> buildKnots(1.,5.,10)
10-element Vector{Float64}:
 0.4285714285714286
 1.0
 1.5714285714285714
 2.142857142857143
 2.7142857142857144
 3.2857142857142856
 3.857142857142857
 4.428571428571429
 5.0
 5.571428571428571

PSplines.buildOm1 — Method

buildOm1(nvertices)

Returns a matrix full of zeros with the appropiate dimensions to apply a selective penalization to the first differences vector. The default 0 value, implies no penalization on the first derivative.

#Examples

julia> buildOm1(5)
4×4 Matrix{Float64}:
 0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0

PSplines.buildOm2 — Method

buildOm2(nvertices)

Returns a matrix full of zeros with the appropiate dimensions to apply a selective penalization to the second differences vector. The default 0 value, implies no penalization on the second derivative.

#Examples

julia> buildOm2(5)
3×3 Matrix{Float64}:
 0.0  0.0  0.0
 0.0  0.0  0.0
 0.0  0.0  0.0

PSplines.buildOm3 — Method

buildOm3(nvertices)

Returns a matrix full of zeros with the appropiate dimensions to apply a selective penalization to the third differences vector. The default 0 value, implies no penalization on the third derivative.

#Examples

julia> buildOm3(5)
2×2 Matrix{Float64}:
 0.0  0.0
 0.0  0.0

PSplines.buildW — Method

buildW(nvertices)

Given the total number of experimental points, build the weight matrix for that number of points. Note that by default all the weights are equal to 1. #Examples

julia> buildOm3(5)
2×2 Matrix{Float64}:
 0.0  0.0
 0.0  0.0

PSplines.equiSpacedPosition — Method

equiSpacedPosition(x, knotsIn)

Returns the bin in which x is contained for the bins defined by the limits in knotsIn. By default if the value is in the limit it is pPosition at the right bin except for the las limit which is positioned to the left. #Examples

julia> equiSpacedPosition(2, [1,2,3,4,5,6,7,8])
2

PSplines.evalBSplineBasis — Method

evalBSplineBasis(x, knots)

Builds evaluation of the cubic BSpline basis corresponding to the knots at x. #Examples

julia> evalBSplineBasis(10.,buildKnots(1.,10., 12))
12-element Vector{Float64}:
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.16666666666666663
 0.6666666666666666
 0.16666666666666666

PSplines.knots — Method

knots(bs::BSpline)

Returns the field knots for a certain BSpline.

PSplines.vertices — Method

vertices(bs::BSpline)

Returns the field vertices for a certain BSpline.