PSplines

Documentation for PSplines.

PSplines.BSpline

PSplines.BSpline2dim

PSplines.D1matrix

PSplines.D2matrix

PSplines.D3matrix

PSplines.Wmatrix

PSplines.buildKnots

PSplines.equiSpacedPosition

PSplines.evalBSpline

PSplines.evalBSplineBasis

PSplines.evalBSplineBasisD1

PSplines.evalBSplineD1

PSplines.expandBSpline

PSplines.knots

PSplines.vertices

PSplines.Ω1matrix

PSplines.Ω2matrix

PSplines.Ω3matrix

PSplines.BSpline — Type

BSpline{T}

bspline is a structures which includes the knots and the vertices of a BSpline.

PSplines.BSpline2dim — Type

BSpline2dim{T}

bspline2dim is a structures which includes the knots in x and y direction and the vertices (in vector form) of a 2 dimensional BSpline.

PSplines.D1matrix — Method

D1matrix(nvertices)

Returns the first difference of dimension nvercies-1 x nvertices. It is important to note that the numerical first derivative of a vector produces a vector with 1 component less than the input vector.

#Examples

julia> D1matrix(3)
2×3 Matrix{Float64}:
 -1.0   1.0  0.0
  0.0  -1.0  1.0

PSplines.D2matrix — Method

D2matrix(nvertices)

Returns the second difference of dimension nvercies-2 x nvertices. It is important to note that the numerical second derivative of a vector produces a vector with 2 component less than the input vector.

#Examples

julia> D2matrix(4)
2×4 Matrix{Float64}:
 1.0  -2.0   1.0  0.0
 0.0   1.0  -2.0  1.0

PSplines.D3matrix — Method

D3matrix(nvertices)

Returns the second difference of dimension nvercies-3 x nvertices. It is important to note that the numerical second derivative of a vector produces a vector with 2 component less than the input vector.

#Examples

julia> D3matrix(5)
2×5 Matrix{Float64}:
 -1.0   3.0  -3.0   1.0  0.0
  0.0  -1.0   3.0  -3.0  1.0

PSplines.Wmatrix — Method

Wmatrix(nvertices)

Given the total number of experimental points, build the weight matrix for that number of points. Note that by default all the weights are equal to 1. #Examples

julia> Wmatrix(5)
2×2 Matrix{Float64}:
 0.0  0.0
 0.0  0.0

PSplines.buildKnots — Method

buildKnots(xin, xend, nvertices)

Returns the knots of a BSpline with beginning xin, end xend with nvertices components. The knots returned by this function are equispaced between xin and xend.

#Examples

julia> buildKnots(1.,5.,10)
10-element Vector{Float64}:
 0.4285714285714286
 1.0
 1.5714285714285714
 2.142857142857143
 2.7142857142857144
 3.2857142857142856
 3.857142857142857
 4.428571428571429
 5.0
 5.571428571428571

PSplines.equiSpacedPosition — Method

equiSpacedPosition(x, knotsIn)

Returns the bin in which x is contained for the bins defined by the limits in knotsIn. By default if the value is in the limit it is pPosition at the right bin except for the las limit which is positioned to the left. #Examples

julia> equiSpacedPosition(2, [1,2,3,4,5,6,7,8])
2

PSplines.evalBSpline — Method

evalBSpline(x, knots, vertices)

Evaluation of the cubic BSpline at x for vertices values vertices and the BSpline basis defined by evalBSplineBasis(x, knots). #Examples

julia> evalBSpline(9.,buildKnots(1.,10., 12), ones(12))
0.9999999999999999

PSplines.evalBSplineBasis — Method

evalBSplineBasis(x, knots)

Builds the evaluation of the cubic BSpline basis corresponding to the knots at x. #Examples

julia> evalBSplineBasis(10.,buildKnots(1.,10., 12))
12-element Vector{Float64}:
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.0
 0.16666666666666663
 0.6666666666666666
 0.16666666666666666

PSplines.evalBSplineBasisD1 — Method

evalBSplineBasisD1(x, knots)

Builds the evaluation of the derivative of the cubic BSpline basis corresponding to the knots at x. #Examples

julia> evalBSplineBasisD1(10.,buildKnots(1.,10., 12))
12-element Vector{Float64}:
  0.0
  0.0
  0.0
  0.0
  ⋮
 -0.5
  0.0
  0.5

PSplines.evalBSplineD1 — Method

evalBSplineD1(x, knots, vertices)

Evaluating the derivative ofthe cubic BSpline at x for vertices values vertices and the BSpline basis derivative defined by evalBSplineBasisD1(x, knots). #Examples

julia> evalBSplineD1(9.,buildKnots(1.,10., 12), ones(12))
0.0

PSplines.expandBSpline — Method

expandBSpline(knots, vertices, xin, xend)

Given the knots and the vertices of a given BSpline with a certain range of validity, it returns a new BSpline which extends the range of validity of the previous BSpline betwenn xin and xend. #Examples

julia> bs2 = expandBSpline(bs1, 0., 5.);
julia> bs2.knots
20-element Vector{Float64}:
 -0.6666666666666663
 -0.33333333333333304
  2.220446049250313e-16
  ⋮
  5.333333333333333
  5.666666666666666
julia> bs2.vertices
20-element Vector{Float64}:
  -1.8150147193435557
  -1.2594147415453065
  -0.7038147637472744
   ⋮
  27.07391859179092
  29.851651947327774

PSplines.knots — Method

knots(bs::BSpline)

Returns the field knots for a certain BSpline.

PSplines.vertices — Method

vertices(bs::BSpline)

Returns the field vertices for a certain BSpline.

PSplines.Ω1matrix — Method

Ω1matrix(nvertices)

Returns a matrix full of zeros with the appropiate dimensions to apply a selective penalization to the first differences vector. The default 0 value, implies no penalization on the first derivative.

#Examples

julia> Ω1matrix(5)
4×4 Matrix{Float64}:
 0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0

PSplines.Ω2matrix — Method

Ω2matrix(nvertices)

Returns a matrix full of zeros with the appropiate dimensions to apply a selective penalization to the second differences vector. The default 0 value, implies no penalization on the second derivative.

#Examples

julia> Ω2matrix(5)
3×3 Matrix{Float64}:
 0.0  0.0  0.0
 0.0  0.0  0.0
 0.0  0.0  0.0

PSplines.Ω3matrix — Method

Ω3matrix(nvertices)

Returns a matrix full of zeros with the appropiate dimensions to apply a selective penalization to the third differences vector. The default 0 value, implies no penalization on the third derivative.

#Examples

julia> Ω3matrix(5)
2×2 Matrix{Float64}:
 0.0  0.0
 0.0  0.0